criacao de arquivos diferentes para commander e funcs
